src/requirements.txt
faker==25.2.0
pyyaml==6.0.1
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.32.1
fastavro==1.9.4
confluent-kafka==2.4.0